validação de cpf e senha adicionados

+++ src/main/java/br/edu/ufersa/sys_scholar/domain/entity/Aluno.java
@@ -12,6 +12,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
+import javax.validation.constraints.Size;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -39,7 +41,7 @@ public class Aluno {
     private String nome;
 
     @Column(unique = true)
-    private Integer cpf;
+    private Long cpf;
 
     // @Column(unique = true)
     // private String usuario;
